penambahan invoicve dan CORS

+++ app/categories/controller.js
@@ -2,6 +2,16 @@ const Category = require("./model");
 
 async function store(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Category")) {
+      // <-- can create Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat kategori`,
+      });
+    }
     // (1) tangkap payload dari _client request_
     let payload = req.body;
     // (2) buat category baru dengan model Category;
@@ -27,6 +37,16 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Category")) {
+      // <-- can update Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate kategori`,
+      });
+    }
     let payload = req.body;
 
     let category = await Category.findOneAndUpdate(
@@ -51,6 +71,16 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Category")) {
+      // <-- can delete Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus kategori`,
+      });
+    }
     // (1) cari dan hapus categori di MongoDB berdasarkan field _id
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
