+++ app/products/controller.js
@@ -6,9 +6,20 @@ const path = require("path");
 const Product = require("./model");
 const Category = require("../products/model");
 const config = require("../config");
+const { policyFor } = require("../policy");
 
 async function store(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Product")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat produk`,
+      });
+    }
+    //-----------------//
     let payload = req.body;
 
     if (payload.category) {
@@ -104,19 +115,31 @@ async function index(req, res, next) {
       criteria = { ...criteria, tags: { $in: tags.map((tag) => tag._id) } };
     }
 
+    let count = await Product.find(criteria).countDocuments(); // <--- perubahan
+
     let products = await Product.find(criteria)
       .limit(parseInt(limit))
       .skip(parseInt(skip))
       .populate("category")
       .populate("tags");
 
-    return res.json(products);
+    return res.json({ data: products, count }); // <--- perubahan
   } catch (err) {
     next(err);
   }
 }
 async function update(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Product")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate produk`,
+      });
+    }
+
     let payload = req.body;
 
     if (payload.category) {
@@ -203,6 +226,17 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Product")) {
+      // <-- can delete
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus produk`,
+      });
+    }
+
     let product = await Product.findOneAndDelete({ _id: req.params.id });
 
     let currentImage = `${config.rootPath}/public/upload/${product.image_url}`;
